add: limit entity and repository(without relation), transaction entity and repository(without relation) edit: structure database and entities

+++ src/main/java/com/test/microservice/repositories/ExpenseCategoryRepository.java
@@ -3,5 +3,5 @@ package com.test.microservice.repositories;
 import com.test.microservice.entities.ExpenseCategory;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ExpenseCategoryRepository extends CrudRepository<ExpenseCategory, Integer> {
+public interface ExpenseCategoryRepository extends CrudRepository<ExpenseCategory, Long> {
 }
